entity depth offset is now a SceneContext
