+++ crates/re_space_view_spatial/src/scene/contexts/mod.rs
@@ -1,0 +1,3 @@
+mod depth_offsets;
+
+pub use depth_offsets::EntityDepthOffsets;
